refactor(context-menu): H.8 — extract useContextMenuKeyboardNav hook
